Move completion wca view tests to their own file
